+++ lib/md_feature/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.2 - 2024-10-07
+
+- Add 'copyWith' method for BLoC data class
+- Rename `map` to `when` in BLoC state and event classes
+- Fix `/bin/sh: fvm dart: command not found` when using fvm
+
 ## 1.0.1 - 2024-09-17
 
 - Add `use_fvm` variable